Change author_create_post functions
We validation and sanitizes the data, then we use isEmpty to check whether there are any errors.

If there are then we render the form again with data by user entered and array of error messages.

If the data is valid then we call isAuthorExists function to search to see if a new Author already exists.

If the new Author already exists we redirect to its detail page.

If not, we save the new Author and redirect to its detail page.