Change bookinstance_create_post functions
We validation and sanitizes the data, then we use isEmpty to check whether there are any errors.

If there are then we call  renderError function to render the form again with data by user entered and array of error messages.

If the data is valid then we call isBookInstanceExists function to search to see if a new BookInstance already exists.

If the new BookInstance already exists we redirect to its detail page.

If not, we save the new BookInstance and redirect to its detail page.