Change author_delete_post method
The author variable obtains the author information of a specific id.

The allBooksByAuthor variable will get all related books on a specific author ID.

If the author variable is not found, you will be redirected to the "author_list" page.

If the author variable is found, check whether the length of the allBooksByAuthor variable is greater than 0. If greater than 0, redirect to the "author_delete" page. If not, it will remove the author.