Change genre_delete_get method
The genre variable obtains the genre information for a specific id.

The allBooksByGenre variable will gets all related books on a specific author ID.

If the genre variable is found, the "genre_delete" page will be rendered.

If the genre variable is not found, you will be redirected to the "genre_list" page.