Change genre_delete_post method
The genre variable obtains the genre information of a specific id.

The allBooksByGenre variable will get all related books on a specific genre ID.

If the genre variable is not found, you will be redirected to the "genre_list" page.

If the genre variable is found, check whether the length of the allBooksByGenre variable is greater than 0. If greater than 0, redirect to the "genre_delete" page. If not, it will remove the genre.