Change book_delete_get method
The book variable obtains the book information for a specific id.

The allBookInstancesByBook variable will gets all related book instances on a specific book ID.

If the book variable is found, the "book_delete" page will be rendered.

If the book variable is not found, you will be redirected to the "book_list" page.