Change book_delete_post method
The book variable obtains the book information of a specific id.

The allBookInstancesByBook variable will get all related book instances on a specific book ID.

If the genre variable is not found, you will be redirected to the "genre_list" page.

If the book variable is found, check whether the length of the allBookInstancesByBook variable is greater than 0. If greater than 0, redirect to the "book_delete" page. If not, it will remove the book.