Change book_update_get method
The book variable obtains the book information for a specific id.

The authors variable obtains the all author information.

The genres variable obtains the all author information.

If the book variable is found, the "book_form" page will be rendered.

If the book variable is not found, you will be redirected to the "book_list" page.