Change author_update_post functions
We validation and sanitizes the data, then we use isEmpty to check whether there are any errors.

If there are then we call  renderError function to render the form again with data by user entered and array of error messages.

If the data is valid then we call isAuthorExists function to search to see if a new Author already exists.

If the new Author already exists we redirect to its detail page.

If not, we update the Author document and redirect to its detail page.